+++ src/ResourceManager/EventHub/Commands.EventHub/Cmdlets/VNetRules/GetAzureEventHubVNetRule.cs
@@ -1,0 +1,81 @@
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using Microsoft.Azure.Commands.EventHub.Models;
+using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+
+namespace Microsoft.Azure.Commands.EventHub.Commands.EventHub
+{    
+    /// <summary>
+    /// 'Get-AzureRmEventHub' Cmdlet gives the details of a / List of EventHub(s)
+    /// <para> If EventHub name provided, a single EventHub detials will be returned</para>
+    /// <para> If EventHub name not provided, list of EventHub will be returned</para>
+    /// </summary>
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "EventHubVNetRule", DefaultParameterSetName = VnetRulePropertiesParameterSet), OutputType(typeof(PSVirtualNetWorkRuleAttributes))]
+    public class GetAzureEventHubVNetRule : AzureEventHubsCmdletBase
+    {
+        [Parameter(Mandatory = true, ParameterSetName = VnetRulePropertiesParameterSet, Position = 0, HelpMessage = "Resource Group Name")]
+        [ResourceGroupCompleter]
+        [ValidateNotNullOrEmpty]
+         public string ResourceGroupName { get; set; }
+
+        [Parameter(Mandatory = true, ParameterSetName = VnetRulePropertiesParameterSet, Position = 1, HelpMessage = "Namespace Name")]
+        [ValidateNotNullOrEmpty]
+        [Alias(AliasNamespaceName)]
+        public string Namespace { get; set; }
+
+        [Parameter(Mandatory = false, ParameterSetName = VnetRulePropertiesParameterSet, Position = 2, HelpMessage = "Virtual Network Rule Name")]
+        public string Name { get; set; }
+
+        [Parameter(Mandatory = true, ParameterSetName = VnetRuleResourceIdParameterSet, ValueFromPipelineByPropertyName = true, Position = 0, HelpMessage = "Virtual Network Rule Resource Id")]
+        [ValidateNotNullOrEmpty]
+        public string ResourceId { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public override void ExecuteCmdlet()
+        {
+            try
+            {
+                if (ParameterSetName.Equals(VnetRuleResourceIdParameterSet))
+                {
+                    LocalResourceIdentifier identifier = new LocalResourceIdentifier(ResourceId);
+                    ResourceGroupName = identifier.ResourceGroupName;
+                    Namespace = identifier.ParentResource;
+                    Name = identifier.ResourceName;
+                }
+
+                if (!string.IsNullOrEmpty(Name))
+                {
+                    // Get a VNet Rule
+                    PSVirtualNetWorkRuleAttributes eventHub = Client.GetVNetRule(ResourceGroupName, Namespace, Name);
+                    WriteObject(eventHub);
+                }
+                else
+                {
+                    IEnumerable<PSVirtualNetWorkRuleAttributes> eventHubsList = Client.ListVNetRule(ResourceGroupName, Namespace);
+                    WriteObject(eventHubsList.ToList(), true);
+                }
+            }
+            catch (Management.EventHub.Models.ErrorResponseException ex)
+            {
+                WriteError(Eventhub.EventHubsClient.WriteErrorforBadrequest(ex));
+            }
+        }
+    }
+}